+++ src/main/java/util/validator/Validator.java
@@ -1,4 +1,4 @@
-package util;
+package util.validator;
 
 import static util.constant.Message.*;
