+++ src/test/java/domain/LottoTest.java
@@ -1,0 +1,20 @@
+package domain;
+
+import java.util.ArrayList;
+import java.util.List;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+public class LottoTest {
+
+    static void
+
+    @Test
+    @DisplayName("로또 랜덤 테스트")
+    void test1() {
+
+        Lotto lotto = new Lotto();
+        List<Integer> lottoList = new ArrayList<>();
+
+    }
+}
